pyproject.toml
[tool.poetry]
name = "podcast-rag"
version = "0.1.0"
description = ""
authors = ["Yals <eliyahu.yalley@long-ai.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
python-dotenv = "1.0.0"
streamlit = "^1.33.0"
streamlit-chat = "^0.1.1"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
pymilvus = "^2.4.0"
scipy = "^1.13.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"